requirements.txt
langchain>=0.2.9
langchain-openai>=0.1.8
langchain-groq>=0.1.5
groq>=0.11.0
openai>=1.43.0
python-dotenv>=1.0.1
tiktoken>=0.7.0
rich>=13.7.1
